common/test-requirements.txt
nose

#Docs
sphinxcontrib-pecanwsme>=0.8
sphinxcontrib-httpdomain
sphinx_rtd_theme
